+++ dataforge-workspace/src/jvmTest/kotlin/space/kscience/dataforge/workspace/FileDataTest.kt
@@ -3,15 +3,16 @@ package space.kscience.dataforge.workspace
 import io.ktor.utils.io.core.Input
 import io.ktor.utils.io.core.Output
 import kotlinx.coroutines.runBlocking
+import space.kscience.dataforge.context.Context
 import space.kscience.dataforge.context.Global
 import space.kscience.dataforge.data.*
-import space.kscience.dataforge.io.IOFormat
-import space.kscience.dataforge.io.io
-import space.kscience.dataforge.io.readUtf8String
-import space.kscience.dataforge.io.writeUtf8String
+import space.kscience.dataforge.io.*
+import space.kscience.dataforge.io.yaml.YamlPlugin
 import space.kscience.dataforge.meta.get
 import space.kscience.dataforge.misc.DFExperimental
 import java.nio.file.Files
+import kotlin.io.path.fileSize
+import kotlin.io.path.toPath
 import kotlin.reflect.KType
 import kotlin.reflect.typeOf
 import kotlin.test.Test
@@ -44,32 +45,38 @@ class FileDataTest {
 
     @Test
     @DFExperimental
-    fun testDataWriteRead() {
-        Global.io.run {
-            val dir = Files.createTempDirectory("df_data_node")
-            runBlocking {
-                writeDataDirectory(dir, dataNode, StringIOFormat)
-                println(dir.toUri().toString())
-                val reconstructed = readDataDirectory(dir) { _, _ -> StringIOFormat }
-                assertEquals(dataNode["dir.a"]?.meta?.get("content"), reconstructed["dir.a"]?.meta?.get("content"))
-                assertEquals(dataNode["b"]?.await(), reconstructed["b"]?.await())
-            }
+    fun testDataWriteRead() = with(Global.io) {
+        val dir = Files.createTempDirectory("df_data_node")
+        runBlocking {
+            writeDataDirectory(dir, dataNode, StringIOFormat)
+            println(dir.toUri().toString())
+            val reconstructed = readDataDirectory(dir) { _, _ -> StringIOFormat }
+            assertEquals(dataNode["dir.a"]?.meta?.get("content"), reconstructed["dir.a"]?.meta?.get("content"))
+            assertEquals(dataNode["b"]?.await(), reconstructed["b"]?.await())
         }
     }
 
 
     @Test
     @DFExperimental
-    fun testZipWriteRead() {
-        Global.io.run {
-            val zip = Files.createTempFile("df_data_node", ".zip")
-            runBlocking {
-                dataNode.writeZip(zip, StringIOFormat)
-                println(zip.toUri().toString())
-                val reconstructed = readDataDirectory(zip) { _, _ -> StringIOFormat }
-                assertEquals(dataNode["dir.a"]?.meta?.get("content"), reconstructed["dir.a"]?.meta?.get("content"))
-                assertEquals(dataNode["b"]?.await(), reconstructed["b"]?.await())
-            }
+    fun testZipWriteRead() = with(Global.io) {
+        val zip = Files.createTempFile("df_data_node", ".zip")
+        runBlocking {
+            dataNode.writeZip(zip, StringIOFormat)
+            println(zip.toUri().toString())
+            val reconstructed = readDataDirectory(zip) { _, _ -> StringIOFormat }
+            assertEquals(dataNode["dir.a"]?.meta?.get("content"), reconstructed["dir.a"]?.meta?.get("content"))
+            assertEquals(dataNode["b"]?.await(), reconstructed["b"]?.await())
         }
+    }
+
+    @Test
+    fun testNonEnvelope() {
+        val context = Context {
+            plugin(YamlPlugin)
+        }
+        val resource = javaClass.classLoader.getResource("SPC.png")!!
+        val data: Envelope = context.io.readEnvelopeFile(resource.toURI().toPath(), true)
+        assertEquals(resource.toURI().toPath().fileSize(), data.data?.size?.toLong())
     }
 }